chore: copy account features from @celo/cryptographic-utils
