+++ package.json
@@ -121,11 +121,14 @@
     "@walletconnect/web3wallet": "^1.14.1",
     "Base64": "^1.3.0",
     "bignumber.js": "^9.1.2",
+    "bip32": "^2.0.6",
+    "bip39": "https://github.com/bitcoinjs/bip39#d8ea080a18b40f301d4e2219a2991cd2417e83c2",
     "clevertap-react-native": "^2.2.1",
     "country-data": "^0.0.31",
     "crypto-js": "^4.2.0",
     "date-fns": "^3.6.0",
     "dot-prop-immutable": "^2.1.1",
+    "ethereumjs-util": "^5.2.0",
     "fast-levenshtein": "^3.0.0",
     "fp-ts": "2.16.9",
     "futoin-hkdf": "^1.5.3",
@@ -139,6 +142,7 @@
     "lodash": "^4.17.21",
     "lottie-react-native": "^5.1.6",
     "node-libs-react-native": "^1.2.1",
+    "randombytes": "^2.0.1",
     "react": "18.3.1",
     "react-async-hook": "^4.0.0",
     "react-i18next": "^14.1.3",
