+++ src/backup/utils.ts
@@ -1,15 +1,15 @@
-import { CELO_DERIVATION_PATH_BASE, generateKeys } from '@celo/cryptographic-utils'
 import CryptoJS from 'crypto-js'
 import { useAsync } from 'react-async-hook'
 import * as bip39 from 'react-native-bip39'
 import { showError } from 'src/alert/actions'
-import { OnboardingEvents } from 'src/analytics/Events'
 import AppAnalytics from 'src/analytics/AppAnalytics'
+import { OnboardingEvents } from 'src/analytics/Events'
 import { ErrorMessages } from 'src/app/ErrorMessages'
 import { getPassword } from 'src/pincode/authentication'
 import { useDispatch, useSelector } from 'src/redux/hooks'
 import { removeStoredItem, retrieveStoredItem, storeItem } from 'src/storage/keychain'
 import Logger from 'src/utils/Logger'
+import { CELO_DERIVATION_PATH_BASE, generateKeys } from 'src/utils/account'
 import { ETHEREUM_DERIVATION_PATH } from 'src/web3/consts'
 import { currentAccountSelector } from 'src/web3/selectors'
